Refactoring for clearer understanding, added a few XXXes
